feat: PostService에 User와의 관계 추가

+++ src/main/java/opensocial/org/community_hub/service/PostService.java
@@ -2,6 +2,7 @@ package opensocial.org.community_hub.service;
 
 import lombok.RequiredArgsConstructor;
 import opensocial.org.community_hub.entity.Post;
+import opensocial.org.community_hub.entity.User;
 import opensocial.org.community_hub.repository.PostRepository;
 import org.springframework.stereotype.Service;
 
@@ -13,20 +14,24 @@ public class PostService {
 
     private final PostRepository postRepository;
 
-    // 게시글 생성
-    public Post createPost(Post post) {
+    public Post createPost(Post post, User user) {
+
+        post.setUser(user); // 게시글에 사용자 정보 추가
         return postRepository.save(post);
     }
 
-    // 게시글 조회
     public Optional<Post> getPostById(Long postId) {
         return postRepository.findById(postId);
     }
 
-    // 게시글 업데이트
-    public Post updatePost(Long postId, Post postDetails) {
+    // 게시글 업데이트 (본인 게시글만 업데이트)
+    public Post updatePost(Long postId, Post postDetails, User user) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found with id " + postId));
+
+        if (!post.getUser().getUserId().equals(user.getUserId())) {
+            throw new RuntimeException("You are not authorized to update this post");
+        }
 
         post.setTitle(postDetails.getTitle());
         post.setContent(postDetails.getContent());
@@ -36,10 +41,15 @@ public class PostService {
         return postRepository.save(post);
     }
 
-    // 게시글 삭제
-    public void deletePost(Long postId) {
+    // 게시글 삭제 (본인 게시글만 삭제)
+    public void deletePost(Long postId, User user) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found with id " + postId));
+
+        if (!post.getUser().getUserId().equals(user.getUserId())) {
+            throw new RuntimeException("You are not authorized to delete this post");
+        }
+
         postRepository.delete(post);
     }
 }
